fix(HyperionEngine.cpp): fix calculation of deltaTime to use SDL_GetTicks() for accurate time tracking feat(HyperionEngine.cpp): update event handling to ensure smoother rendering and improve user experience feat(HyperionEngine.h): change data types of _lastUpdate, _currentTick, and _deltaTime to uint64_t and Uint32 for consistency and efficiency feat(main.cpp): change vsync option to false for better performance feat(TestScreen.cpp): increase time threshold for character generation to improve visual effect and user experience

+++ src/include/engine/HyperionEngine.h
@@ -28,10 +28,9 @@ namespace hyperion::engine {
 
         std::vector<screen::ScreenConsole *> _screenConsoles;
 
-        Uint64 _lastUpdate = 0;
-        Uint64 _currentTick = 0;
-        double _deltaTime = 0;
-
+        uint64_t _lastUpdate = 0;
+        uint64_t _currentTick = 0;
+        Uint32 _deltaTime = 0;
 
 
         bool _running;
